+++ cosmohub/api/security/token.py
@@ -1,3 +1,7 @@
+import calendar
+
+from datetime import datetime, timedelta
+
 from flask import current_app
 from flask_restful import marshal
 from itsdangerous import TimedJSONWebSignatureSerializer
@@ -24,7 +28,18 @@ class Token(object):
         )
     
     def dump(self):
+        # FIXME: HACK: Manually add 'exp' claim in payload to account for
+        # https://github.com/pallets/itsdangerous/issues/73
+        exp = calendar.timegm(
+            (
+                datetime.utcnow() + timedelta(
+                    seconds=current_app.config['TOKEN_EXPIRES_IN_DEFAULT']
+                )
+            ).timetuple()
+        )
+        
         token = {
+            'exp' : exp,
             'user' : marshal(self._user, fields.UserToken),
             'privilege' : self._privilege.attr,
         }