+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased][unreleased]
+### Fixed
+- Manually add 'exp' to token payload to fix `itsdangerous` misbehaviour. (Pau Tallada)
 
 
 ## [2.2.1] - 2016-10-19